+++ frameworks/cocos2d-x/cocos/2d/CCAnimationCache.cpp
@@ -100,7 +100,7 @@ void AnimationCache::parseVersion1(const ValueMap& animations)
             continue;
         }
 
-        ssize_t frameNameSize = frameNames.size();
+        unsigned long frameNameSize = frameNames.size();
         Vector<AnimationFrame*> frames(frameNameSize);
 
         for (auto& frameName : frameNames)
