+++ frameworks/cocos2d-x/cocos/2d/CCSpriteBatchNode.cpp
@@ -46,7 +46,7 @@ NS_CC_BEGIN
 * creation with Texture2D
 */
 
-SpriteBatchNode* SpriteBatchNode::createWithTexture(Texture2D* tex, ssize_t capacity/* = DEFAULT_CAPACITY*/)
+SpriteBatchNode* SpriteBatchNode::createWithTexture(Texture2D* tex, unsigned long capacity/* = DEFAULT_CAPACITY*/)
 {
     SpriteBatchNode *batchNode = new (std::nothrow) SpriteBatchNode();
     if(batchNode && batchNode->initWithTexture(tex, capacity))
@@ -63,7 +63,7 @@ SpriteBatchNode* SpriteBatchNode::createWithTexture(Texture2D* tex, ssize_t capa
 * creation with File Image
 */
 
-SpriteBatchNode* SpriteBatchNode::create(const std::string& fileImage, ssize_t capacity/* = DEFAULT_CAPACITY*/)
+SpriteBatchNode* SpriteBatchNode::create(const std::string& fileImage, unsigned long capacity/* = DEFAULT_CAPACITY*/)
 {
     SpriteBatchNode *batchNode = new (std::nothrow) SpriteBatchNode();
     if(batchNode && batchNode->initWithFile(fileImage, capacity))
@@ -79,7 +79,7 @@ SpriteBatchNode* SpriteBatchNode::create(const std::string& fileImage, ssize_t c
 /*
 * init with Texture2D
 */
-bool SpriteBatchNode::initWithTexture(Texture2D *tex, ssize_t capacity/* = DEFAULT_CAPACITY*/)
+bool SpriteBatchNode::initWithTexture(Texture2D *tex, unsigned long capacity/* = DEFAULT_CAPACITY*/)
 {
     if(tex == nullptr)
     {
@@ -187,7 +187,7 @@ bool SpriteBatchNode::init()
 /*
 * init with FileImage
 */
-bool SpriteBatchNode::initWithFile(const std::string& fileImage, ssize_t capacity/* = DEFAULT_CAPACITY*/)
+bool SpriteBatchNode::initWithFile(const std::string& fileImage, unsigned long capacity/* = DEFAULT_CAPACITY*/)
 {
     Texture2D *texture2D = Director::getInstance()->getTextureCache()->addImage(fileImage);
     return initWithTexture(texture2D, capacity);
@@ -303,7 +303,7 @@ void SpriteBatchNode::removeChild(Node *child, bool cleanup)
     Node::removeChild(sprite, cleanup);
 }
 
-void SpriteBatchNode::removeChildAtIndex(ssize_t index, bool doCleanup)
+void SpriteBatchNode::removeChildAtIndex(unsigned long index, bool doCleanup)
 {
     CCASSERT(index>=0 && index < _children.size(), "Invalid index");
     removeChild(_children.at(index), doCleanup);
@@ -338,7 +338,7 @@ void SpriteBatchNode::sortAllChildren()
                 child->sortAllChildren();
             }
 
-            ssize_t index=0;
+            unsigned long index=0;
 
             //fast dispatch, give every child a new atlasIndex based on their relative zOrder (keep parent -> child relations intact)
             // and at the same time reorder descendants and the quads to the right index
@@ -352,12 +352,12 @@ void SpriteBatchNode::sortAllChildren()
     }
 }
 
-void SpriteBatchNode::updateAtlasIndex(Sprite* sprite, ssize_t* curIndex)
+void SpriteBatchNode::updateAtlasIndex(Sprite* sprite, unsigned long* curIndex)
 {
     auto& array = sprite->getChildren();
     auto count = array.size();
     
-    ssize_t oldIndex = 0;
+    unsigned long oldIndex = 0;
 
     if( count == 0 )
     {
@@ -414,7 +414,7 @@ void SpriteBatchNode::updateAtlasIndex(Sprite* sprite, ssize_t* curIndex)
     }
 }
 
-void SpriteBatchNode::swap(ssize_t oldIndex, ssize_t newIndex)
+void SpriteBatchNode::swap(unsigned long oldIndex, unsigned long newIndex)
 {
     CCASSERT(oldIndex>=0 && oldIndex < (int)_descendants.size() && newIndex >=0 && newIndex < (int)_descendants.size(), "Invalid index");
 
@@ -464,7 +464,7 @@ void SpriteBatchNode::increaseAtlasCapacity()
     // if we're going beyond the current TextureAtlas's capacity,
     // all the previously initialized sprites will need to redo their texture coords
     // this is likely computationally expensive
-    ssize_t quantity = (_textureAtlas->getCapacity() + 1) * 4 / 3;
+    unsigned long quantity = (_textureAtlas->getCapacity() + 1) * 4 / 3;
 
     CCLOG("cocos2d: SpriteBatchNode: resizing TextureAtlas capacity from [%d] to [%d].",
         static_cast<int>(_textureAtlas->getCapacity()),
@@ -478,7 +478,7 @@ void SpriteBatchNode::increaseAtlasCapacity()
     }
 }
 
-void SpriteBatchNode::reserveCapacity(ssize_t newCapacity)
+void SpriteBatchNode::reserveCapacity(unsigned long newCapacity)
 {
     if (newCapacity <= _textureAtlas->getCapacity())
         return;
@@ -491,7 +491,7 @@ void SpriteBatchNode::reserveCapacity(ssize_t newCapacity)
     }
 }
 
-ssize_t SpriteBatchNode::rebuildIndexInOrder(Sprite *parent, ssize_t index)
+unsigned long SpriteBatchNode::rebuildIndexInOrder(Sprite *parent, unsigned long index)
 {
     CCASSERT(index>=0 && index < _children.size(), "Invalid index");
 
@@ -522,7 +522,7 @@ ssize_t SpriteBatchNode::rebuildIndexInOrder(Sprite *parent, ssize_t index)
     return index;
 }
 
-ssize_t SpriteBatchNode::highestAtlasIndexInChild(Sprite *sprite)
+unsigned long SpriteBatchNode::highestAtlasIndexInChild(Sprite *sprite)
 {
     auto& children = sprite->getChildren();
 
@@ -536,7 +536,7 @@ ssize_t SpriteBatchNode::highestAtlasIndexInChild(Sprite *sprite)
     }
 }
 
-ssize_t SpriteBatchNode::lowestAtlasIndexInChild(Sprite *sprite)
+unsigned long SpriteBatchNode::lowestAtlasIndexInChild(Sprite *sprite)
 {
     auto& children = sprite->getChildren();
 
@@ -550,7 +550,7 @@ ssize_t SpriteBatchNode::lowestAtlasIndexInChild(Sprite *sprite)
     }
 }
 
-ssize_t SpriteBatchNode::atlasIndexForChild(Sprite *sprite, int nZ)
+unsigned long SpriteBatchNode::atlasIndexForChild(Sprite *sprite, int nZ)
 {
     auto& siblings = sprite->getParent()->getChildren();
     auto childIndex = siblings.getIndex(sprite);
@@ -720,7 +720,7 @@ void SpriteBatchNode::setTexture(Texture2D *texture)
 // SpriteSheet Extension
 //implementation SpriteSheet (TMXTiledMapExtension)
 
-void SpriteBatchNode::insertQuadFromSprite(Sprite *sprite, ssize_t index)
+void SpriteBatchNode::insertQuadFromSprite(Sprite *sprite, unsigned long index)
 {
     CCASSERT( sprite != nullptr, "Argument must be non-nullptr");
     CCASSERT( dynamic_cast<Sprite*>(sprite), "CCSpriteBatchNode only supports Sprites as children");
@@ -745,7 +745,7 @@ void SpriteBatchNode::insertQuadFromSprite(Sprite *sprite, ssize_t index)
     sprite->updateTransform();
 }
 
-void SpriteBatchNode::updateQuadFromSprite(Sprite *sprite, ssize_t index)
+void SpriteBatchNode::updateQuadFromSprite(Sprite *sprite, unsigned long index)
 {
     CCASSERT(sprite != nullptr, "Argument must be non-nil");
     CCASSERT(dynamic_cast<Sprite*>(sprite) != nullptr, "CCSpriteBatchNode only supports Sprites as children");
