+++ frameworks/cocos2d-x/cocos/2d/CCTMXXMLParser.cpp
@@ -824,9 +824,9 @@ void TMXMapInfo::endElement(void* /*ctx*/, const char *name)
                 unsigned char *deflated = nullptr;
                 Size s = layer->_layerSize;
                 // int sizeHint = s.width * s.height * sizeof(uint32_t);
-                ssize_t sizeHint = s.width * s.height * sizeof(unsigned int);
+                unsigned long sizeHint = s.width * s.height * sizeof(unsigned int);
                 
-                ssize_t CC_UNUSED inflatedLen = ZipUtils::inflateMemoryWithHint(buffer, len, &deflated, sizeHint);
+                unsigned long CC_UNUSED inflatedLen = ZipUtils::inflateMemoryWithHint(buffer, len, &deflated, sizeHint);
                 CCASSERT(inflatedLen == sizeHint, "inflatedLen should be equal to sizeHint!");
                 
                 free(buffer);
