+++ frameworks/cocos2d-x/cocos/3d/CCAABB.h
@@ -108,7 +108,7 @@ public:
     /**
      * update the _min and _max from the given point.
      */
-    void updateMinMax(const Vec3* point, ssize_t num);
+    void updateMinMax(const Vec3* point, unsigned long num);
     
     /**
      * Transforms the bounding box by the given transformation matrix.
